refactor: split sim.rs, add determinism test, polish docs
Structural: src/sim.rs shrinks from 3373 to 1186 lines by extracting
three cohesive submodules that preserve the public API unchanged:

- sim/construction.rs -- new, new_with_hooks, validation, topology
  builders, from_parts, dispatch/reposition setters, hook registration
- sim/topology.rs -- runtime add/remove/reassign for lines, elevators,
  stops, and groups plus topology queries and shortest_route
- sim/lifecycle.rs -- reroute, settle, despawn, disable/enable, route
  invalidation, population queries, per-elevator metrics, service mode

Tests + examples:

- tests/deterministic_replay.rs proves the crate's determinism promise
  end-to-end: two identical 5000-tick runs with ETD dispatch must yield
  byte-identical event streams and metrics, with a paired sanity test
  confirming the fixture exercises real boarding and delivery.
- examples/space_elevator.rs loads assets/config/space_elevator.ron and
  delivers five riders across the 1000-unit tether, demonstrating
  'stops at arbitrary positions' on extreme geometry.

Docs:

- ARCHITECTURE.md section 3 phase numbering fixed (AdvanceQueue is
  phase 4; Doors/Loading/Metrics renumbered to 6/7/8).
- lib.rs and README.md point at ARCHITECTURE.md as the canonical
  long-form source to prevent drift.
- Scan / LOOK / ETD module docs cite source literature (Denning 1967,
  Merten 1970, Barney and dos Santos 1985).

Polish:

- src/graph/ empty directory removed.
- register_extensions! doctest converted from ignore to runnable.
- package.metadata.docs.rs all-features = true added so the energy
  feature renders on docs.rs.

Pre-existing fixes bundled in:

- examples/basic.rs panicked on startup because .stop() appends to the
  builder's default stops; switched to .stops() which replaces.
- src/topology.rs test module annotated with the clippy allows it needs.
- Six benchmark ElevatorConfig literals updated to include
  restricted_stops, service_mode, inspection_speed_factor, and the
  feature-gated energy_profile so cargo bench compiles again.
