+++ README.md
@@ -170,6 +170,10 @@ component accessors, per-entity extension storage for game-specific data, and a
 query builder for filtering and iterating entities by component composition.
 Lifecycle hooks let you inject custom logic before or after any phase.
 
+For per-phase semantics (events emitted, edge cases, design rationale), see
+[`ARCHITECTURE.md`](crates/elevator-core/ARCHITECTURE.md) — the single
+canonical reference.
+
 ## Dispatch Strategies
 
 Four built-in strategies ship with the crate. All implement the `DispatchStrategy` trait.
