+++ crates/elevator-core/ARCHITECTURE.md
@@ -156,7 +156,20 @@ Groups without a registered strategy skip this phase entirely.
 
 **Events:** `ElevatorRepositioning`
 
-### Phase 4: Movement (`systems/movement.rs`)
+### Phase 4: AdvanceQueue (`systems/advance_queue.rs`)
+
+Reconciles each elevator's phase and target stop with the front of its
+[`DestinationQueue`]. When imperative callers have pushed a stop via
+`push_destination` or `push_destination_front`, this phase redirects the
+car before movement is applied. If the front of the queue matches the
+current target, the phase is a no-op.
+
+Queue entries are consumed when a loading cycle completes at the target
+stop, so imperative and dispatch-driven itineraries compose naturally.
+
+**Events:** `ElevatorAssigned` (when a new target is adopted from the queue)
+
+### Phase 5: Movement (`systems/movement.rs`)
 
 Applies trapezoidal velocity profile physics (via `movement::tick_movement`)
 to all elevators in `MovingToStop` phase. The profile has three regions:
@@ -175,7 +188,7 @@ emit `ElevatorRepositioned` instead of `ElevatorArrived`.
 **Key design:** Physics parameters (max_speed, acceleration, deceleration) are
 per-elevator, stored on the `Elevator` component.
 
-### Phase 5: Doors (`systems/doors.rs`)
+### Phase 6: Doors (`systems/doors.rs`)
 
 Ticks the `DoorState` finite-state machine for each elevator:
 
@@ -190,7 +203,7 @@ Phase transitions on completion:
 
 **Events:** `DoorOpened`, `DoorClosed`
 
-### Phase 6: Loading (`systems/loading.rs`)
+### Phase 7: Loading (`systems/loading.rs`)
 
 Boards and exits riders at elevators in `Loading` phase. Uses a **two-pass
 read-then-write** approach to avoid aliasing issues:
@@ -206,7 +219,7 @@ Boarding checks weight capacity and rider preferences; failures emit
 
 **Events:** `RiderBoarded`, `RiderExited`, `RiderRejected`
 
-### Phase 7: Metrics (`systems/metrics.rs`)
+### Phase 8: Metrics (`systems/metrics.rs`)
 
 Reads events emitted during the current tick (via `EventBus::peek()`) and
 updates aggregate `Metrics`:
