+++ crates/elevator-core/Cargo.toml
@@ -10,6 +10,10 @@ rust-version.workspace = true
 categories = ["game-engines", "simulation"]
 keywords = ["elevator", "dispatch", "simulation", "ecs", "building"]
 
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
 [features]
 default = ["traffic"]
 traffic = ["dep:rand"]
