+++ crates/elevator-core/benches/multi_line_bench.rs
@@ -83,6 +83,11 @@ fn multi_group_config(
                         starting_stop: serves[0],
                         door_open_ticks: 5,
                         door_transition_ticks: 3,
+                        restricted_stops: Vec::new(),
+                        #[cfg(feature = "energy")]
+                        energy_profile: None,
+                        service_mode: None,
+                        inspection_speed_factor: 0.25,
                     }
                 })
                 .collect();
@@ -157,6 +162,11 @@ fn single_group_config(num_stops: u32, num_elevators: u32) -> SimConfig {
             starting_stop: StopId(i % num_stops),
             door_open_ticks: 5,
             door_transition_ticks: 3,
+            restricted_stops: Vec::new(),
+            #[cfg(feature = "energy")]
+            energy_profile: None,
+            service_mode: None,
+            inspection_speed_factor: 0.25,
         })
         .collect();
