+++ docs/src/dispatch.md
@@ -15,7 +15,7 @@ Direction indicators (`going_up`/`going_down`) are derived automatically from ea
 
 ## Imperative dispatch (destination queue)
 
-If you want saga-style "just tell the elevator where to go" control, you don't need to implement a `DispatchStrategy`. Every elevator carries a `DestinationQueue` (a FIFO of stop `EntityId`s) that you can push to directly:
+If you just want to tell an elevator where to go — no decision-making strategy required — every elevator carries a `DestinationQueue` (a FIFO of stop `EntityId`s) that you can push to directly:
 
 ```rust,no_run
 # use elevator_core::prelude::*;
@@ -24,7 +24,7 @@ If you want saga-style "just tell the elevator where to go" control, you don't n
 # let stop_a: EntityId = todo!();
 # let stop_b: EntityId = todo!();
 sim.push_destination(elev, stop_a).unwrap();        // enqueue at back
-sim.push_destination_front(elev, stop_b).unwrap();  // saga's forceNow
+sim.push_destination_front(elev, stop_b).unwrap();  // jump ahead of the queue
 sim.clear_destinations(elev).unwrap();              // cancel pending work
 let queue: &[EntityId] = sim.destination_queue(elev).unwrap();
 ```
