+++ docs/src/metrics-and-events.md
@@ -173,7 +173,7 @@ println!("Total moves:       {}", m.total_moves());
 | `total_settled()` | Cumulative riders settled as residents |
 | `total_rerouted()` | Cumulative riders rerouted from resident phase |
 | `total_distance()` | Sum of all elevator travel distance |
-| `total_moves()` | Total rounded-floor transitions across all elevators (passing-floor crossings + arrivals; analogous to elevator-saga's `moveCount`) |
+| `total_moves()` | Total rounded-floor transitions across all elevators (passing-floor crossings + arrivals) |
 | `utilization_by_group()` | Per-group fraction of elevators currently moving |
 | `avg_utilization()` | Average utilization across all groups |
 | `reposition_distance()` | Total elevator distance traveled while repositioning |
